third_party/blink/web_tests/wpt_internal/app-history/navigate-event/navigate-appHistory-back-same-document.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
async_test(t => {
  // Wait for after the load event so that the navigation doesn't get converted
  // into a replace navigation.
  window.onload = () => t.step_timeout(() => {
    appHistory.navigate("#foo").then(() => {
      appHistory.onnavigate = t.step_func_done(e => {
        assert_true(e.cancelable);
        assert_true(e.canRespond);
        assert_false(e.userInitiated);
        assert_true(e.hashChange);
        assert_equals(new URL(e.destination.url).hash, "");
        assert_true(e.destination.sameDocument);
        assert_equals(e.formData, null);
        assert_equals(e.info, "hi");
      });
      assert_true(appHistory.canGoBack);
      appHistory.back({ navigateInfo: "hi" });
    });
  }, 0);
}, "navigate event for appHistory.back() - same-document");
</script>
